requirements.txt
chardet==3.0.4
git+git://github.com/gogaz/clashroyale.git@master#egg=clashroyale
codecov
Django==3.0.3
djangorestframework==3.11.0
flake8
gunicorn==20.0.4
psycopg2-binary
python-box
requests
django-command-log==0.2.2
celery
redis
django-celery-results
django-celery-beat